trpc code/file restructure
