+++ src/app/api/trpc/[trpc]/route.ts
@@ -2,7 +2,7 @@ import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 import { type NextRequest } from "next/server";
 
 import { env } from "@/env";
-import { appRouter } from "@/server/api/root";
+import { appRouter } from "@/server/api/routers/_app";
 import { createTRPCContext } from "@/server/api/trpc";
 
 /**
