+++ src/trpc/TRPCReactProvider.tsx
@@ -1,0 +1,46 @@
+"use client";
+
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { loggerLink, unstable_httpBatchStreamLink } from "@trpc/client";
+import { useState } from "react";
+import SuperJSON from "superjson";
+
+import { trpc } from "@/trpc/client";
+import { getBaseUrl } from "@/trpc/utils";
+
+export default function TRPCReactProvider({
+  children,
+  cookies,
+}: {
+  children: React.ReactNode;
+  cookies: string;
+}) {
+  const [queryClient] = useState(() => new QueryClient({}));
+  const [trpcClient] = useState(() =>
+    trpc.createClient({
+      transformer: SuperJSON,
+      links: [
+        loggerLink({
+          enabled: (op) =>
+            process.env.NODE_ENV === "development" ||
+            (op.direction === "down" && op.result instanceof Error),
+        }),
+        unstable_httpBatchStreamLink({
+          url: getBaseUrl() + "/api/trpc",
+          headers: () => {
+            return {
+              cookie: cookies,
+              "x-trpc-source": "react",
+            };
+          },
+        }),
+      ],
+    })
+  );
+
+  return (
+    <trpc.Provider client={trpcClient} queryClient={queryClient}>
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    </trpc.Provider>
+  );
+}
