+++ src/trpc/client.ts
@@ -1,0 +1,11 @@
+"use client";
+
+import "client-only";
+import { createTRPCReact } from "@trpc/react-query";
+import { type inferRouterInputs, type inferRouterOutputs } from "@trpc/server";
+import { type AppRouter } from "@/server/api/routers/_app";
+
+export const trpc = createTRPCReact<AppRouter>();
+
+export type RouterInputs = inferRouterInputs<AppRouter>;
+export type RouterOutputs = inferRouterOutputs<AppRouter>;
